fixed linux compile errors

<==== Mercurial data ====>
Branch: dev_3.0.0_gui

+++ nx_cloud/libcloud_db/src/settings.cpp
@@ -219,7 +219,7 @@ void Settings::loadConfiguration()
     m_accountManager.passwordResetCodeExpirationTimeout =
         std::chrono::seconds(m_settings.value(
             kPasswordResetCodeExpirationTimeout,
-            kDefaultPasswordResetCodeExpirationTimeout.count()).toInt());
+            (qlonglong)kDefaultPasswordResetCodeExpirationTimeout.count()).toInt());
 
     //auth
     m_auth.rulesXmlPath = m_settings.value(kAuthXmlPath, kDefaultAuthXmlPath).toString();
